Introduce SynchronizedImage API.
This allows Android clients to supply a release callback along with
external images, which permits copy-free synchronization. It also paves
the way for supporting Camera with Vulkan. This API is not necessary for
Metal because of Metal's shared ownership semantics.

The OpenGL driver implements this by storing a list of in-use
SynchronizedImage objects (typically there are only 0 or 1 of these) and
querying their fences during beginFrame. If an image is no longer
needed, then its callback is scheduled to be triggered in a manner
similar to the BufferDescriptor callbacks.

This API has been tested with a fork of our `hello-camera` Kotlin
sample, which we will add to the repo soon.